# docs/openapi-banco-orbital.yaml
...
openapi: 3.0.3
info:
  title: Banco Orbital — API
  version: 1.0.0-draft
  description: |
    API del home banking de **Banco Orbital**. Este documento es el contrato entre
    nuestro backend y nuestro frontend, y la referencia para los otros bancos del
    curso que quieran comparar.

    ## Alcance de la v1

    Cuentas, tarjetas, préstamos y plazos fijos. **Seguros, CEDEARs y reservas
    quedan para la v2** — decisión tomada para que el alcance entre en el tiempo
    disponible.

    Los proveedores externos simulados (recargas de celular y empresas de
    servicios) **no están acá**: viven en una API aparte, `banco-proveedores`, que
    este banco consume por HTTP como si fuera un tercero real.

    ## Convenciones

    | Tema | Convención |
    |---|---|
    | Autenticación | JWT de Clerk en `Authorization: Bearer <token>` |
    | Errores | `{ "error": "mensaje" }`, con `details` opcional. Igual que el Banco Central |
    | Paginación | `page` y `limit` por query; respuesta `{ page, limit, count, data }` |
    | Importes | `number` con 2 decimales. Internamente se calculan en decimal exacto |
    | Nombres | Español, según `docs/GLOSARIO.md`. Campos en `snake_case`, como ya los devuelven los DTOs |
    | Idempotencia | Header `Idempotency-Key` (UUID v4) en **todo endpoint que mueve plata** |

    ### Sobre `monto` contra `importe`

    Dentro de nuestro banco el campo siempre se llama **`monto`**. `importe` aparece
...
    contrato. No mezclarlos.

    ### Sobre las rutas
...
servers:
...
security:
  - ClerkJWT: []
...
components:
  securitySchemes:
    ClerkJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token de sesión emitido por Clerk. El backend lo valida y resuelve el
        `usuario` y su `persona` asociada. Todos los endpoints bajo `/api` lo
        requieren, además de usuario activo y perfil completo.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, format: uuid }
      description: |
        UUID v4. Protege contra el doble envío: reintentar con la misma clave
        devuelve la respuesta original en vez de ejecutar la operación otra vez.
      example: 3f2504e0-4f89-41d3-9a0c-0305e82c3301
  responses:
    NoAutorizado:
      description: Falta el token, es inválido, o el usuario está inactivo.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 'No autorizado.' }
    Prohibido:
      description: El usuario no tiene permiso, o el recurso pertenece a otra persona.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 'No puedes operar sobre una cuenta que no te pertenece.' }
    NoEncontrado:
      description: El recurso no existe.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 'No existe el recurso solicitado.' }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Mensaje legible. En producción los 5xx se reemplazan por uno genérico.
          example: 'Saldo insuficiente.'
        details:
          type: object
          nullable: true
          description: |
            Detalle adicional. Presente en errores de validación (con `fieldErrors`)
            y en desarrollo. Nunca expone internos de la base en producción.
      additionalProperties: false
...
    Monto:
      type: number
      format: double
      description: |
        Importe con 2 decimales. Se almacena como `NUMERIC(18,2)` y se calcula en
        decimal exacto; se expone como número por comodidad del cliente, que sólo
        lo formatea para mostrar.
      example: 184500.20
...
    Tarjeta:
      type: object
      properties:
        id: { type: string, format: uuid }
        persona_id: { type: string, format: uuid }
        tipo:
          type: string
          enum: [debito, credito]
          example: credito
        numero_enmascarado:
          type: string
          description: Sólo los últimos 4 dígitos. El número completo no se expone nunca.
          example: '**** **** **** 4821'
        cuenta_id:
          type: string
          format: uuid
          nullable: true
          description: Cuenta asociada. Obligatoria en débito, `null` en crédito.
        limite:
          allOf: [{ $ref: '#/components/schemas/Monto' }]
          nullable: true
          description: Sólo en crédito.
        disponible:
          allOf: [{ $ref: '#/components/schemas/Monto' }]
          nullable: true
          description: 'Límite menos el consumo del período. Sólo en crédito.'
        estado:
          type: string
          enum: [activa, bloqueada, vencida]
          example: activa
        vencimiento: { type: string, example: '09/30' }
        created_at: { type: string, format: date-time }
    Autorizacion:
      type: object
      properties:
        id: { type: string, format: uuid }
        tarjeta_id: { type: string, format: uuid }
        comercio: { type: string, example: Supermercado Norte }
        monto: { $ref: '#/components/schemas/Monto' }
        cuotas: { type: integer, minimum: 1, example: 3 }
        estado:
          type: string
          enum: [aprobada, rechazada]
          example: aprobada
        motivo_rechazo:
          type: string
          nullable: true
          description: Presente sólo si fue rechazada.
          example: Límite disponible insuficiente.
        created_at: { type: string, format: date-time }
...
paths:
...
  /api/tarjetas:
    post:
      tags: [Tarjetas]
      summary: Emitir una tarjeta
      description: |
        Emite una tarjeta de débito, atada a una cuenta, o de crédito, con límite
        propio. El número completo no se devuelve nunca: sólo el enmascarado.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [persona_id, tipo]
              properties:
                persona_id: { type: string, format: uuid }
                tipo: { type: string, enum: [debito, credito] }
                cuenta_id:
                  type: string
                  format: uuid
                  nullable: true
                  description: Obligatorio si el tipo es `debito`.
                limite:
                  allOf: [{ $ref: '#/components/schemas/Monto' }]
                  nullable: true
                  description: Obligatorio si el tipo es `credito`.
            examples:
              debito:
                value: { persona_id: 9a1b2c3d-0000-4000-8000-000000000010, tipo: debito, cuenta_id: 4b1c8e2a-0000-4000-8000-000000000001 }
              credito:
                value: { persona_id: 9a1b2c3d-0000-4000-8000-000000000010, tipo: credito, limite: 500000 }
      responses:
        '201':
          description: Tarjeta emitida.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tarjeta' }
        '400':
          description: Falta `cuenta_id` en débito, o `limite` en crédito.
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '403': { $ref: '#/components/responses/Prohibido' }
  /api/personas/{personaId}/tarjetas:
    get:
      tags: [Tarjetas]
      summary: Listar las tarjetas de una persona
      parameters:
        - name: personaId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Tarjetas del titular.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Tarjeta' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '403': { $ref: '#/components/responses/Prohibido' }
  /api/tarjetas/{tarjetaId}/autorizaciones:
    post:
      tags: [Tarjetas]
      summary: Autorizar un consumo
      description: |
        Simula la autorización en tiempo real de una compra. En débito valida el saldo
        de la cuenta asociada; en crédito, el límite disponible. Siempre queda
        registrada, aprobada o rechazada.
      parameters:
        - name: tarjetaId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [comercio, monto]
              properties:
                comercio: { type: string, example: Supermercado Norte }
                monto: { $ref: '#/components/schemas/Monto' }
                cuotas: { type: integer, minimum: 1, default: 1, example: 3 }
      responses:
        '201':
          description: Consumo aprobado.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Autorizacion' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
        '409':
          description: La tarjeta está bloqueada o vencida.
        '422':
          description: Saldo o límite insuficiente. La autorización queda registrada como rechazada.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Autorizacion' }
              example:
                estado: rechazada
                motivo_rechazo: Límite disponible insuficiente.
                monto: 90000
  /api/tarjetas/{tarjetaId}/bloqueo:
    post:
      tags: [Tarjetas]
      summary: Bloquear o desbloquear una tarjeta
      description: |
        Cambia el estado de la tarjeta. Es la acción típica de "perdí la tarjeta".

        **El asistente con IA no puede ejecutar esto**: es de solo lectura y deriva
        al cliente a la app o a un humano.
      parameters:
        - name: tarjetaId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [accion]
              properties:
                accion: { type: string, enum: [bloquear, desbloquear] }
      responses:
        '200':
          description: Estado actualizado.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tarjeta' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '403': { $ref: '#/components/responses/Prohibido' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
  /api/tarjetas/{tarjetaId}/resumen:
    get:
      tags: [Tarjetas]
      summary: Resumen de cuenta
      description: Sólo para tarjetas de crédito.
      parameters:
        - name: tarjetaId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: periodo
          in: query
          required: false
          schema: { type: string, example: '2026-09' }
          description: Mes en formato `YYYY-MM`. Por defecto, el actual.
      responses:
        '200':
          description: Resumen del período.
          content:
            application/json:
              schema:
                type: object
                properties:
                  periodo: { type: string, example: '2026-09' }
                  vencimiento: { type: string, format: date, example: '2026-10-10' }
                  total_a_pagar: { $ref: '#/components/schemas/Monto' }
                  pago_minimo: { $ref: '#/components/schemas/Monto' }
                  cft: { type: number, example: 101.22 }
                  consumos:
                    type: array
                    items: { $ref: '#/components/schemas/Autorizacion' }
        '400':
          description: La tarjeta es de débito y no tiene resumen.
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '404': { $ref: '#/components/responses/NoEncontrado' }

  # ───────────────────────────── Préstamos ─────────────────────────────
...